shl_recommendation_system/requirements.txt
# SHL Assessment Recommendation System
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.29.0
pydantic>=2.5.0

# Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank_bm25>=0.2.2

# LLM
google-generativeai>=0.3.0

# Data & Web
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0

# Scraper (JS-rendered catalog)
playwright>=1.40.0

# Dev & Quality
mypy>=1.8.0
ruff>=0.1.0
pytest>=7.4.0
